fix: renders artworksList in two columns
