+++ pwa-nextjs/app/globals.css
@@ -3,14 +3,14 @@
 @tailwind utilities;
 
 :root {
-  --background: #ffffff;
+  --background: #626262;
   --foreground: #171717;
 }
 
 @media (prefers-color-scheme: dark) {
   :root {
     --background: #0a0a0a;
-    --foreground: #ededed;
+    --foreground: #9a3838;
   }
 }
 
@@ -23,7 +23,7 @@ body {
 .search-box{
   width: 100%;
   background: #fff;
-  margin: 200px auto 0;
+  margin: 20px auto 0;
   border-radius: 5px;
   border: solid black 1px
 }
@@ -36,11 +36,11 @@ body {
 
 input{
   flex: 1;
-  height: 50px;
+  height: 30px;
   background: transparent;
   border: 0;
   outline: 0;
-  font-size: 18px;
+  font-size: 12px;
   color: #333
 }
